renamed pages for better clarity
